feat: respect activated channel

+++ src/db/manager.rs
@@ -9,8 +9,8 @@ impl DbManager {
 
     pub async fn set_active_channel(
         &self,
-        guild_id: i64,
-        channel_id: i64,
+        guild_id: u64,
+        channel_id: u64,
     ) -> Result<(), anyhow::Error> {
         sqlx::query(
             "INSERT INTO guild_channels (guild_id, channel_id)
@@ -18,18 +18,18 @@ impl DbManager {
              ON CONFLICT (guild_id)
              DO UPDATE SET channel_id = $2",
         )
-        .bind(guild_id)
-        .bind(channel_id)
+        .bind(guild_id as i64)
+        .bind(channel_id as i64)
         .execute(&self.pool)
         .await?;
 
         Ok(())
     }
 
-    pub async fn get_active_channel(&self, guild_id: i64) -> Result<Option<i64>, anyhow::Error> {
+    pub async fn get_active_channel(&self, guild_id: u64) -> Result<Option<i64>, anyhow::Error> {
         let channel_id =
             sqlx::query_scalar("SELECT channel_id FROM guild_channels WHERE guild_id = $1")
-                .bind(guild_id)
+                .bind(guild_id as i64)
                 .fetch_optional(&self.pool)
                 .await?;
 
@@ -38,16 +38,16 @@ impl DbManager {
 
     pub async fn add_compliment(
         &self,
-        guild_id: i64,
-        user_id: i64,
+        guild_id: u64,
+        user_id: u64,
         compliment: i64,
     ) -> Result<(), anyhow::Error> {
         sqlx::query(
             "INSERT INTO user_compliments (guild_id, user_id, compliment)
              VALUES ($1, $2, $3)",
         )
-        .bind(guild_id)
-        .bind(user_id)
+        .bind(guild_id as i64)
+        .bind(user_id as i64)
         .bind(compliment)
         .execute(&self.pool)
         .await?;
@@ -57,16 +57,16 @@ impl DbManager {
 
     pub async fn get_user_compliments(
         &self,
-        guild_id: i64,
-        user_id: i64,
+        guild_id: u64,
+        user_id: u64,
     ) -> Result<Vec<i64>, anyhow::Error> {
         let compliments = sqlx::query_scalar(
             "SELECT compliment
              FROM user_compliments
              WHERE guild_id = $1 AND user_id = $2",
         )
-        .bind(guild_id)
-        .bind(user_id)
+        .bind(guild_id as i64)
+        .bind(user_id as i64)
         .fetch_all(&self.pool)
         .await?;
